requirements.txt
requests
beautifulsoup4
colored
progressbar2